+++ CHANGELOG.md
@@ -5,6 +5,13 @@
 ### Added
 
 - `Innmind\TimeContinuum\Earth\PointInTime\Now::elapsedSince()` now use the high resolution time api when comparing to another instance of `Now` (to avoid errors when time _goes back_ due to the system clock re-synchronisation)
+- `Innmind\TimeContinuum\Earth\Period\Year::of()`
+- `Innmind\TimeContinuum\Earth\Period\Second::of()`
+- `Innmind\TimeContinuum\Earth\Period\Month::of()`
+- `Innmind\TimeContinuum\Earth\Period\Minute::of()`
+- `Innmind\TimeContinuum\Earth\Period\Millisecond::of()`
+- `Innmind\TimeContinuum\Earth\Period\Hour::of()`
+- `Innmind\TimeContinuum\Earth\Period\Day::of()`
 
 ### Changed